Split AgentProcessor: extract LLM request + swappable response-format components
Zero behavior change. agent-llm-request.ts owns the in-flight slot,
transport attempts, chunk streaming, and atomic settle appends;
agent-response-format.ts owns the pure content -> outcome strategy
(fenced-ts moved verbatim, corrective feedback included in the outcome).
The processor stays the orchestrator and keeps every append's idempotency
and blocking discipline; emitted bodies and keys are byte-identical, and
the existing test suites pass unchanged.
